+++ src/line.rs
@@ -28,14 +28,18 @@ impl<'src> Line<'src> {
 
   pub(crate) fn is_quiet(&self) -> bool {
     match self.fragments.first() {
-      Some(Fragment::Text { token }) => token.lexeme().starts_with('@'),
+      Some(Fragment::Text { token }) => {
+        token.lexeme().starts_with('@') || token.lexeme().starts_with("-@")
+      }
       _ => false,
     }
   }
 
-  pub(crate) fn is_infallable(&self) -> bool {
+  pub(crate) fn is_infallible(&self) -> bool {
     match self.fragments.first() {
-      Some(Fragment::Text { token }) => token.lexeme().starts_with('-'),
+      Some(Fragment::Text { token }) => {
+        token.lexeme().starts_with('-') || token.lexeme().starts_with("@-")
+      }
       _ => false,
     }
   }