+++ src/parser.rs
@@ -60,7 +60,7 @@ impl<'tokens, 'src> Parser<'tokens, 'src> {
       expected: self
         .expected
         .iter()
-        .cloned()
+        .copied()
         .filter(|kind| *kind != ByteOrderMark)
         .collect::<Vec<TokenKind>>(),
       found: self.next()?.kind,
@@ -77,7 +77,7 @@ impl<'tokens, 'src> Parser<'tokens, 'src> {
   fn rest(&self) -> impl Iterator<Item = Token<'src>> + 'tokens {
     self.tokens[self.next..]
       .iter()
-      .cloned()
+      .copied()
       .filter(|token| token.kind != Whitespace)
   }
 
@@ -654,10 +654,10 @@ impl<'tokens, 'src> Parser<'tokens, 'src> {
     };
 
     Ok(Parameter {
-      name,
-      kind,
       default,
       export,
+      kind,
+      name,
     })
   }
 
@@ -729,7 +729,13 @@ impl<'tokens, 'src> Parser<'tokens, 'src> {
     let name = Name::from_identifier(self.presume(Identifier)?);
     let lexeme = name.lexeme();
 
-    if Keyword::DotenvLoad == lexeme {
+    if Keyword::AllowDuplicateRecipes == lexeme {
+      let value = self.parse_set_bool()?;
+      return Ok(Set {
+        value: Setting::AllowDuplicateRecipes(value),
+        name,
+      });
+    } else if Keyword::DotenvLoad == lexeme {
       let value = self.parse_set_bool()?;
       return Ok(Set {
         value: Setting::DotenvLoad(value),
@@ -745,6 +751,12 @@ impl<'tokens, 'src> Parser<'tokens, 'src> {
       let value = self.parse_set_bool()?;
       return Ok(Set {
         value: Setting::PositionalArguments(value),
+        name,
+      });
+    } else if Keyword::WindowsPowershell == lexeme {
+      let value = self.parse_set_bool()?;
+      return Ok(Set {
+        value: Setting::WindowsPowerShell(value),
         name,
       });
     }
@@ -771,7 +783,7 @@ impl<'tokens, 'src> Parser<'tokens, 'src> {
       self.expect(BracketR)?;
 
       Ok(Set {
-        value: Setting::Shell(setting::Shell { command, arguments }),
+        value: Setting::Shell(setting::Shell { arguments, command }),
         name,
       })
     } else {
@@ -1709,6 +1721,12 @@ mod tests {
   }
 
   test! {
+    name: set_allow_duplicate_recipes_implicit,
+    text: "set allow-duplicate-recipes",
+    tree: (justfile (set allow_duplicate_recipes true)),
+  }
+
+  test! {
     name: set_dotenv_load_true,
     text: "set dotenv-load := true",
     tree: (justfile (set dotenv_load true)),
@@ -1772,6 +1790,24 @@ mod tests {
     name: set_shell_with_two_arguments,
     text: "set shell := ['bash', '-cu', '-l']",
     tree: (justfile (set shell "bash" "-cu" "-l")),
+  }
+
+  test! {
+    name: set_windows_powershell_implicit,
+    text: "set windows-powershell",
+    tree: (justfile (set windows_powershell true)),
+  }
+
+  test! {
+    name: set_windows_powershell_true,
+    text: "set windows-powershell := true",
+    tree: (justfile (set windows_powershell true)),
+  }
+
+  test! {
+    name: set_windows_powershell_false,
+    text: "set windows-powershell := false",
+    tree: (justfile (set windows_powershell false)),
   }
 
   test! {