+++ tests/json.rs
@@ -39,10 +39,12 @@ fn alias() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -65,10 +67,12 @@ fn assignment() {
       "first": null,
       "recipes": {},
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -104,10 +108,12 @@ fn body() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -153,10 +159,12 @@ fn dependencies() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -239,10 +247,64 @@ fn dependency_argument() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
+      },
+      "warnings": [],
+    }),
+  );
+}
+
+#[test]
+fn duplicate_recipes() {
+  test(
+    "
+      set allow-duplicate-recipes
+      alias f := foo
+
+      foo:
+      foo bar:
+    ",
+    json!({
+      "first": "foo",
+      "aliases": {
+        "f": {
+          "name": "f",
+          "target": "foo",
+        }
+      },
+      "assignments": {},
+      "recipes": {
+        "foo": {
+          "body": [],
+          "dependencies": [],
+          "doc": null,
+          "name": "foo",
+          "parameters": [
+            {
+              "name": "bar",
+              "export": false,
+              "default": null,
+              "kind": "singular",
+            },
+          ],
+          "priors": 0,
+          "private": false,
+          "quiet": false,
+          "shebang": false,
+        }
+      },
+      "settings": {
+        "allow_duplicate_recipes": true,
+        "dotenv_load": null,
+        "export": false,
+        "positional_arguments": false,
+        "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -271,10 +333,12 @@ fn doc_comment() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -291,10 +355,12 @@ fn empty_justfile() {
       "first": null,
       "recipes": {},
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -420,10 +486,12 @@ fn parameters() {
         },
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -488,10 +556,12 @@ fn priors() {
         },
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -520,10 +590,12 @@ fn private() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -552,10 +624,12 @@ fn quiet() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -602,6 +676,7 @@ fn settings() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": true,
         "export": true,
         "positional_arguments": true,
@@ -609,6 +684,7 @@ fn settings() {
           "arguments": ["b", "c"],
           "command": "a",
         },
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -640,10 +716,12 @@ fn shebang() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),
@@ -672,10 +750,12 @@ fn simple() {
         }
       },
       "settings": {
+        "allow_duplicate_recipes": false,
         "dotenv_load": null,
         "export": false,
         "positional_arguments": false,
         "shell": null,
+        "windows_powershell": false,
       },
       "warnings": [],
     }),